Refactor main form state logic
